# Conflicts:
#	src/api/src/main/kotlin/com/tenmm/tilserver/draft/adapter/inbound/controller/GetDraftController.kt
#	src/api/src/main/kotlin/com/tenmm/tilserver/draft/adapter/inbound/controller/SaveDraftController.kt
#	src/api/src/main/kotlin/com/tenmm/tilserver/draft/adapter/inbound/controller/SyncDraftController.kt
#	src/api/src/main/kotlin/com/tenmm/tilserver/draft/adapter/inbound/controller/model/GetDraftResponse.kt

+++ src/api/src/main/kotlin/com/tenmm/tilserver/post/application/inbound/model/PostSaveConfirmResult.kt
@@ -4,5 +4,5 @@ import com.tenmm.tilserver.common.domain.OperationResult
 
 data class PostSaveConfirmResult(
     val operationResult: OperationResult,
-    val monthlyPublishCount: Long,
+    val monthlyPublishCount: Int,
 )
